Refactor and migrate native library build structure
Moved native notification library to a new "linuxlib" structure with updated build scripts and CMake configuration. Replaced executable creation with shared library generation and adjusted references to match the new layout. Removed unused main.c and updated project settings to align with the changes.
