+++ .idea/artifacts/composeApp_jvm_1_0_0.xml
@@ -1,8 +1,6 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="composeApp-jvm-1.0.0">
     <output-path>$PROJECT_DIR$/demo/composeApp/build/libs</output-path>
-    <root id="archive" name="composeApp-jvm-1.0.0.jar">
-      <element id="module-output" name="K-Notify.demo.composeApp.jvmMain" />
-    </root>
+    <root id="archive" name="composeApp-jvm-1.0.0.jar" />
   </artifact>
 </component>
