+++ linuxlib/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.10)
 
-project(NotificationExample C)
+project(notification C)
 
 set(CMAKE_C_STANDARD 11)
 
@@ -10,8 +10,8 @@ pkg_check_modules(LIBNOTIFY REQUIRED libnotify)
 include_directories(${LIBNOTIFY_INCLUDE_DIRS})
 link_directories(${LIBNOTIFY_LIBRARY_DIRS})
 
-add_executable(NotificationExample main.c
-        notification_library.c
-        notification_library.h)
+add_library(notification SHARED
+        linux_notification_library.c
+        linux_notification_library.h)
 
-target_link_libraries(NotificationExample ${LIBNOTIFY_LIBRARIES})
+target_link_libraries(notification ${LIBNOTIFY_LIBRARIES})
